requirements.txt
langchain
langchain-community
langchain-anthropic
anthropic
faiss-cpu
pypdf
python-dotenv
sentence-transformers